app/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0

openai==2.7.2
langchain==1.0.5
langchain-openai==1.0.2
langchain-community==0.4.1
langchain-classic==1.0.0
langchain-core==1.0.4
langchain-text-splitters==1.0.0
langchain-experimental==0.4.0

chromadb==1.3.4
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0

python-dotenv==1.2.1

posthog==5.4.0
tiktoken==0.12.0
tenacity==9.1.2
